feat: added deletion cleanup

+++ src/app/api/v1/website/route.ts
@@ -105,7 +105,9 @@ export async function PATCH(request: Request) {
 
       for (const field of bucketFiles) {
         if (website?.[field] && existingData[0]?.[field] !== website[field]) {
-          deleteObject(existingData[0][field])
+          if (existingData[0]?.[field]) {
+            deleteObject(existingData[0][field])
+          }
         }
       }
     }
